Implement basic /portfolio Route and styling
