+++ src/components/activityCard.js
@@ -1,5 +1,6 @@
 import React from "react";
 import "./activityCard.css";
+import { ActivitiesList } from "./activitiesList";
 
 export class ActivityCard extends React.Component {
   render() {
@@ -12,21 +13,4 @@ export class ActivityCard extends React.Component {
       </div>
     );
   }
-}
-
-/*Component returns a list of activities*/
-export function ActivitiesList(props) {
-  const activities = props.activities;
-  const activitiesList = activities.map((activity, i) => (
-    <li key={"activity_" + i}>
-      <div>
-        <h4>{activity.title}</h4>
-        <h5>{activity.institution}</h5>
-        <h5>{activity.dates}</h5>
-        <h5>{activity.location}</h5>
-        <p>{activity.description}</p>
-      </div>
-    </li>
-  ));
-  return <ul> {activitiesList} </ul>;
 }
