+++ src/components/portfolioCard.css
@@ -1,0 +1,108 @@
+/*Styling for mobile screens*/
+
+.portfolio-ul {
+  list-style: none;
+  padding: 10px;
+  margin: 10px 0;
+  background-color: #f4f6f6;
+  border: solid 1px white;
+  border-radius: 8px;
+}
+
+.portfolio-li {
+  background-color: white;
+  padding: 15px 5px;
+  margin: 20px 5px;
+  border: solid 1px white;
+  border-radius: 8px;
+}
+
+.portfolio h4 {
+  margin: 5px 5px 15px 5px;
+}
+
+.portfolio h5 {
+  margin: 0;
+  font-weight: normal;
+  color: grey;
+}
+
+.portfolio p {
+  margin: 10px 0 0 0;
+  padding: 5px;
+}
+
+.stack ul {
+  list-style: none;
+  padding: 0 5px;
+}
+
+.stack li {
+  display: inline-block;
+  padding: 8px 12px;
+  margin: 5px;
+  background-color: #d0ece7;
+  border-radius: 10px;
+}
+
+.project-links {
+  margin: 20px 0 10px 0;
+}
+
+.project-links button {
+  margin: 0 10px;
+  display: inline-block;
+  padding: 8px 12px;
+  border-radius: 10px;
+  background-color: white;
+  color: black;
+  border: 1px solid #808f8c; /* Green */
+  transition-duration: 0.4s;
+  font-family: -apple-system, BlinkMacSystemFont, "Segoe UI", "Roboto", "Oxygen",
+    "Ubuntu", "Cantarell", "Fira Sans", "Droid Sans", "Helvetica Neue",
+    sans-serif;
+}
+
+.project-links button:hover {
+  background-color: #808f8c; /* Green */
+  color: white;
+  cursor: pointer;
+}
+
+/*Styling for desktop screens*/
+@media screen and (min-width: 800px) {
+  .portfolio {
+    max-width: 1240px;
+    margin: auto;
+  }
+
+  .portfolio-ul {
+    padding: 20px;
+    margin: 20px 0;
+    display: flex;
+    flex-direction: row;
+    justify-content: space-evenly;
+    flex-wrap: wrap;
+  }
+
+  .portfolio-li {
+    max-width: 500px;
+    padding: 20px 10px;
+    margin: 30px;
+    display: flex;
+    flex-direction: column;
+  }
+
+  .project-info {
+    flex: 1;
+  }
+
+  .project-links button {
+    margin: 0 20px;
+  }
+
+  .portfolio p {
+    margin: 10px 0 0 0;
+    padding: 5px 15px;
+  }
+}
